[PORT] secure_uninstall to v8
