--- secure_uninstall/README.md
@@ -1,5 +1,0 @@
-Secure Uninstall
-================
-
-Ask Master Password ('admin_passwd' key from config file)
-before to proceed to module uninstallation
