+++ secure_uninstall/__init__.py
@@ -1,1 +1,1 @@
-from . import module
+from . import models
