+++ secure_uninstall/i18n/fr.po
@@ -1,0 +1,88 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* secure_uninstall
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 8.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2016-06-09 09:44+0000\n"
+"PO-Revision-Date: 2016-06-09 11:52+0200\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Plural-Forms: \n"
+"Language: fr\n"
+"X-Generator: Poedit 1.8.7.1\n"
+
+#. module: secure_uninstall
+#: help:uninstall.check.wizard,password:0
+msgid "'admin_passwd' value from Odoo configuration file (aka 'Master Password')"
+msgstr "la valeur 'admin_passwd' du fichier de configuration Odoo (aussi appelé 'Mot de passe Maître')"
+
+#. module: secure_uninstall
+#: view:uninstall.check.wizard:secure_uninstall.view_uninstall_wizard_form
+msgid "Apply"
+msgstr "Appliquer"
+
+#. module: secure_uninstall
+#: view:uninstall.check.wizard:secure_uninstall.view_uninstall_wizard_form
+msgid "Cancel"
+msgstr "Annuler"
+
+#. module: secure_uninstall
+#: view:uninstall.check.wizard:secure_uninstall.view_uninstall_wizard_form
+msgid "Check"
+msgstr "Vérifier"
+
+#. module: secure_uninstall
+#: field:uninstall.check.wizard,create_uid:0
+msgid "Created by"
+msgstr "Créé par"
+
+#. module: secure_uninstall
+#: field:uninstall.check.wizard,create_date:0
+msgid "Created on"
+msgstr "Créé le"
+
+#. module: secure_uninstall
+#: field:uninstall.check.wizard,id:0
+msgid "ID"
+msgstr "ID"
+
+#. module: secure_uninstall
+#: view:uninstall.check.wizard:secure_uninstall.view_uninstall_wizard_form
+msgid "If you want uninstall module, write required password."
+msgstr "Si vous voulez désinstaller un module, écrivez le mot de passe requis."
+
+#. module: secure_uninstall
+#: field:uninstall.check.wizard,write_uid:0
+msgid "Last Updated by"
+msgstr "Mise à jour par"
+
+#. module: secure_uninstall
+#: field:uninstall.check.wizard,write_date:0
+msgid "Last Updated on"
+msgstr "Mis à jour"
+
+#. module: secure_uninstall
+#: model:ir.model,name:secure_uninstall.model_ir_module_module
+msgid "Module"
+msgstr "Module"
+
+#. module: secure_uninstall
+#: field:uninstall.check.wizard,password:0
+msgid "Password"
+msgstr "Mot de passe"
+
+#. module: secure_uninstall
+#: view:uninstall.check.wizard:secure_uninstall.view_uninstall_wizard_form
+msgid "Some columns can be dropped and there is no way to recover that except restoring a database backup"
+msgstr "Certaines colonnes peuvent être supprimé et il n'y a aucun moyen de les retrouver excepté en restaurant une sauvegarde de la base de données"
+
+#. module: secure_uninstall
+#: view:uninstall.check.wizard:secure_uninstall.view_uninstall_wizard_form
+msgid "Uninstall module process remove all data managed by the module."
+msgstr "Le processus de désinstallation supprime toutes les données gérées par le module."
