+++ secure_uninstall/i18n/secure_uninstall.pot
@@ -1,0 +1,87 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* secure_uninstall
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 8.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2016-06-09 09:44+0000\n"
+"PO-Revision-Date: 2016-06-09 09:44+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: secure_uninstall
+#: help:uninstall.check.wizard,password:0
+msgid "'admin_passwd' value from Odoo configuration file (aka 'Master Password')"
+msgstr ""
+
+#. module: secure_uninstall
+#: view:uninstall.check.wizard:secure_uninstall.view_uninstall_wizard_form
+msgid "Apply"
+msgstr ""
+
+#. module: secure_uninstall
+#: view:uninstall.check.wizard:secure_uninstall.view_uninstall_wizard_form
+msgid "Cancel"
+msgstr ""
+
+#. module: secure_uninstall
+#: view:uninstall.check.wizard:secure_uninstall.view_uninstall_wizard_form
+msgid "Check"
+msgstr ""
+
+#. module: secure_uninstall
+#: field:uninstall.check.wizard,create_uid:0
+msgid "Created by"
+msgstr ""
+
+#. module: secure_uninstall
+#: field:uninstall.check.wizard,create_date:0
+msgid "Created on"
+msgstr ""
+
+#. module: secure_uninstall
+#: field:uninstall.check.wizard,id:0
+msgid "ID"
+msgstr ""
+
+#. module: secure_uninstall
+#: view:uninstall.check.wizard:secure_uninstall.view_uninstall_wizard_form
+msgid "If you want uninstall module, write required password."
+msgstr ""
+
+#. module: secure_uninstall
+#: field:uninstall.check.wizard,write_uid:0
+msgid "Last Updated by"
+msgstr ""
+
+#. module: secure_uninstall
+#: field:uninstall.check.wizard,write_date:0
+msgid "Last Updated on"
+msgstr ""
+
+#. module: secure_uninstall
+#: model:ir.model,name:secure_uninstall.model_ir_module_module
+msgid "Module"
+msgstr ""
+
+#. module: secure_uninstall
+#: field:uninstall.check.wizard,password:0
+msgid "Password"
+msgstr ""
+
+#. module: secure_uninstall
+#: view:uninstall.check.wizard:secure_uninstall.view_uninstall_wizard_form
+msgid "Some columns can be dropped and there is no way to recover that except restoring a database backup"
+msgstr ""
+
+#. module: secure_uninstall
+#: view:uninstall.check.wizard:secure_uninstall.view_uninstall_wizard_form
+msgid "Uninstall module process remove all data managed by the module."
+msgstr ""
+
