+++ secure_uninstall/views/wizard_view.xml
